Trying out things

+++ .local/bin/updatedotfiles.sh
@@ -1,5 +1,6 @@
 #!/bin/sh
 
-git clone -q https://github.com/ifananvity/dotfiles.git
-mv -f dotfiles/.* $HOME
-rm -rf dotfiles
+git clone -q https://github.com/ifananvity/dotfiles.git > /dev/null 2>&1
+for dot in $(ls dotfiles/ -a); do
+	[ ! $dot = ".." ] && [ ! $dot = "." ] && mv -f dotfiles/$dot $HOME
+done && rm -rf dotfiles/
